Work in progress

+++ src/entityprops.hpp
@@ -12,8 +12,8 @@ namespace framework {
 
 struct entityprops {
   uint64_t id{};
-  uint32_t frame{};
-  uint32_t last_frame{};
+  uint64_t frame{};
+  uint64_t last_frame{};
   double_t angle{};
   uint8_t alpha{255};
   bool visible{true};
